requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
nptyping==1.4.4
    # via -r requirements.in
numpy==1.21.5
    # via
    #   -r requirements.in
    #   nptyping
pillow==8.4.0
    # via -r requirements.in
typish==1.9.3
    # via nptyping
